requirements.txt
# Core AI and API dependencies
google-genai
python-dotenv
requests

# Image processing
Pillow
opencv-python

# Video processing and editing
moviepy
imageio-ffmpeg

# Audio processing
pydub

# Data processing
numpy
pandas

# Web framework
fastapi
uvicorn[standard]

# Database with PostgreSQL
sqlalchemy
psycopg2-binary
alembic

# Async database support
asyncpg

# Social media APIs
aiohttp
facebook-sdk
facebook-business
python-instagram

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# File handling
aiofiles


# Monitoring & Logging
prometheus-client

# Development tools
pytest
pytest-asyncio
httpx